pyproject.toml
[project]
name = "game-analytics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "clickhouse-connect>=0.14.0",
    "flask>=3.1.3",
    "kafka-python-ng",
]
